Make it easier for tools/bazel wrapper script to find bazel-real
The launcher script uses "exec -a" to launch the wrapper, but if
the wrapper script is itself a Bash script, Bash will set "$0" to be
the path to the wrapper script, not the path to the launcher script.

To work around this, we've been having our wrapper script search the
user's PATH environment variable for bazel-real, but that doesn't
work well with the IntelliJ plugin for Bazel, which may not use the
expected PATH environment.

--
Change-Id: I402ad29d5b809be8e687e217e19c03e7ac3eb972
Reviewed-on: https://cr.bazel.build/7550
PiperOrigin-RevId: 140851992
MOS_MIGRATED_REVID=140851992