Fix NPE in RClassGenerator when dev has code in unnamed package
Just put the R class in the unnamed package.

If we disable the RClassGenerator and use the
old code path, we would have constructed a cmdline
flag set to "null" for aapt, and aapt would write
"package null;" which also wouldn't work...

This may allow you to build an android_library
but it may be hard to build an android_binary
because ultimately AAPT will reject an empty
"package" attribute in the AndroidManifest.xml
when you try to build the APK, so you'd need
to stamp that differently.

--
PiperOrigin-RevId: 140945125
MOS_MIGRATED_REVID=140945125