Rollback of commit 23a254c3f326273fff9849033f3542ab64188264.
*** Reason for rollback ***

Memory regression: []

*** Original change description ***

Pass direct jars to Jack.

The R class generator produces a compiled jar, not a java source file.
This means all android libraries/binaries with resources will need to
convert this jar to jack to ensure that they have access to jars.

This change will not fix Jack builds, but it is a necessary step toward
fixing them.

--
PiperOrigin-RevId: 141091215
MOS_MIGRATED_REVID=141091215