Fix test configured target creation code that was dropping split deps.
Short story: Dependency -> BuildConfiguration maps can have multiple values
because of split transitions.

This is unfortunately one of those instances where the test logic is forked from
production. So this doesn't actually demonstrate bugs in production code. We
already have tests in ConfigurationsForTargetsTest that directly check the
production logic. So this cl's primary value is to unbreak tests that depend on
the forked logic.

--
PiperOrigin-RevId: 141094830
MOS_MIGRATED_REVID=141094830