Windows, JNI: implement nativeGetLongPath
Implement a JNI method that can resolve 8dot3
style paths. This is necessary because we need to
be able to compare 8dot3 style paths and long
paths [1], and because implementing this correctly
in Java seems to be impossible [2].

This change also adds tests for the JNI isJunction
implementation.

See https://github.com/bazelbuild/bazel/issues/2101

[1] https://github.com/bazelbuild/bazel/issues/2145
[2] https://github.com/bazelbuild/bazel/issues/2145#issuecomment-266766716

--
PiperOrigin-RevId: 142123750
MOS_MIGRATED_REVID=142123750