Pulls in two changes from github for the dd_plist library.
https://github.com/3breadt/dd-plist/commit/3fa1fa56c82df169e06079feb54ba281ad41cae6
which stabilizes the ordering of items in the plist and is equivalent to what happens in Apple's reference code.

https://github.com/3breadt/dd-plist/commit/a5fb37384d4ca7a3b5039a7605ca0e9ca926a6ec
which outputs ids in the plist equivalent to Apple's reference code.

We still differ from Apple's code in that we de-dupe booleans, arrays, and dictionaries which produces smaller binaries. Via testing I have determined that the tooling that was having trouble with our "bad" plists, appears to be happy with our output if we continue our de-duping so I left it in there. Fixing it so that it was exactly equivalent to Apple's output would be a significantly larger engineering effort.

--
PiperOrigin-RevId: 142166474
MOS_MIGRATED_REVID=142166474