Fix cycle in dependency graph during bootstrap
This cycle was due to JacocoCoverage_deploy.jar being an implicit dependency
of java_binary. We do not need the binary during bootstrap because we are not
running with coverage but we need the target. Simply move it to exists as a
java_import in tools/jdk.

Fixes #2269.
To be cherry-picked for #2246.

--
Change-Id: I63aee0e76b305833ac0c0c2cae5abe1cbad5e6d1
Reviewed-on: https://cr.bazel.build/8036
PiperOrigin-RevId: 142537244
MOS_MIGRATED_REVID=142537244