Bazel client; implement NormalizePath
This method can normalize paths with "." and ".."
and multiple "/" characters.

E.g. normalize("../foo/./bar/../baz") = "foo/baz"

This method enables us implementing PathExists on
Windows. If the path to check is too long, we need
to prefix it with "\\?\" for the Windows API
functions to work, but then the path must be
fully normalized and in Windows format. We already
have functions to convert a path to Windows format
but that doesn't normalize; with this function we
can finally convert paths like "/c/foo/../bar" to
L"\\?\c:\foo" and check if it exists.

See https://github.com/bazelbuild/bazel/issues/2107
See https://github.com/bazelbuild/bazel/issues/2181

--
PiperOrigin-RevId: 142648194
MOS_MIGRATED_REVID=142648194