Load JNI library on Windows from the runfiles tree as a fallback
This is the behavior of the UnixJniLoader. In order to do that, this
change also moved the runfiles support on Windows in its own library
that the WindowsJniLoader can use to load the JNI library from tests.

Also add the JNI library on Windows to all test that use the JNI library
on Unix.

Fixes #2300.

--
Change-Id: I2eb9207c3aa310d95912a48f64f676957c47cd34
Reviewed-on: https://cr.bazel.build/8045
PiperOrigin-RevId: 143114397
MOS_MIGRATED_REVID=143114397