Bazel client, Windows: fix MSYS root retrieval
The file_windows.cc:MsysRoot::Get() function was
incorrectly assuming that bash.exe is usually
under c:\tools\msys64\bin\bash.exe when in reality
it's c:\tools\msys64\usr\bin\bash.exe, so we need
to walk up one more directory.

Unfortunately the test was assuming the same so
the bug wasn't caught.

The new code is a bit more foolproof in that it
looks for an msys-looking final path segment and
assumes that to be the msys root.

Fixes https://github.com/bazelbuild/bazel/issues/2323

--
PiperOrigin-RevId: 143372193
MOS_MIGRATED_REVID=143372193