Refactor JavaSkylarkApiProvider to source its info from passed-in providers instead of getting them from the base.
This is a preparatory change to allow aspects to bind JavaSkylarkApiProvider with a different name.

java_proto_library needs to be able to advertise its jars from each aspect. The elegant way to do this would be to bind the "java" skylark provider. Unfortunately, proto_library also advertises this provider, even though it is incorrect to do so in the context of java_proto_library, because the jars advertised by the base configured target are certainly not going to be produced.

Unless we can override providers when the merged configured target is produced, we then need to be able to bind a java provider with a different name. Unfortunately the implementation of this accesses providers from the configured target, which again belong to the proto_library, not the aspect. This CL allows the java skylark api provider to serve data defined at construction time rather than from hard-coded providers from the configured target.

--
PiperOrigin-RevId: 143445977
MOS_MIGRATED_REVID=143445977