java_common.create_provider now takes depsets ) rjars = java_common.create_provider(     compile_time_jars = depset(),     runtime_jars = merged_runtime.transitive_runtime_deps, )
This avoids linearizing the runtime_deps (with the corresponding memory
issues). Must be a JavaProvider for proper interaction with native rules
but cannot just be a simple merge since runtime_deps should not
contribute to compile of the dependent rules.

Note, this will effect a change of the already-released API; however,
function marked as undocumented in an experimental object....

Change-Id: I54542a5d57c75e762b2276e0a1988816901a0def
PiperOrigin-RevId: 155384266