Refactor ResourceContainer: make top level and use AutoValue Builder support.
This merges the AndroidResourceContainerBuilder (which it's not even clear is related to
the nested ResourceContainer!) into the newly generated ResourceContainer.Builder.

It also seemed ridiculous for ResourceContainer to get so large and still be subordinate
to AndroidResourcesProvider, especially when it's getting passed around in a lot of other
places (look how many imports needed fixing!). This CL makes it its own top level class.

This allows for easy modification of an existing instance: call toBuilder on it, set
the properties you want set, and then call build.

--
PiperOrigin-RevId: 143574468
MOS_MIGRATED_REVID=143574468