Release scripts: add a git note with the cherrypick origin
If the patch-id differ because of merge conflict or subtle adjacent
change, the cherrypick origin was missing, resulting in non existent
cherrypick lines in the changelog.

If the cherrypicks are done by the script, adds the cherrypick origin
to a git note and use that origin when generating the release note. Also
fallback to the commit hash of the resulting commit if the cherrypick
origin cannot be found.

Fixes #2272.

--
Change-Id: I311bbe777ea82be05f9fba0a658bab075ed1bd97
Reviewed-on: https://cr.bazel.build/8042
PiperOrigin-RevId: 144075867
MOS_MIGRATED_REVID=144075867