Jarjar the third-party dependencies of the Jacoco test runner, except on Windows. This way, we don't need a new release to make coverage collection happen.
This is the fifth try on commit 883e2dd80189931cc43cf1b23817b90b88969687 . This time, it's disabled on Windows because two bugs (#2306 and #2342) collude to make it impossible to run Java binaries during the build on Windows and jarjar is a Java binary.

Tested by building //src:bazel on a Windows machine with --output_base=<something with a $ sign in it>. I also verified that the genrule gets run by adding an "exit 1" at its beginning.

Fifth time is the charm! Hopefully we don't get to seven.

--
PiperOrigin-RevId: 144818587
MOS_MIGRATED_REVID=144818587