Fix bootstrapping in Docker images.
Turns out, we couldn't run jarjar because the Java launcher script looks for the .jars in the runfiles and build-runfiles is stubbed out during bootstrapping.

The only reason why this worked at all is that sandboxing *also* doesn't work during bootstrapping but it causes the creation of symlinks that happened to be just in the right place for the Java launcher to find the .jars .

The fix is:

  - Explicitly disable sandboxing during bootstrapping so that coincidences like this don't happen again
  - Pass a --javabase and --host_javabase option during the bootstrap build so that we don't need any symlinks to access to JVM
  - Invoke jarjar using its deploy jar instead of the launcher script.

That was fun.

--
PiperOrigin-RevId: 145083357
MOS_MIGRATED_REVID=145083357