Move the test environment setup to the TestRunnerAction.
This is part of unifying the test strategies; here, I'm working towards
sharing the test environment setup.

This change moves code from previously closed source code into the open
source parts. The internal change is not visible, making this look like an
addition rather than a move..

The next change hooks this up to Bazel. I did this in order to reduce the
size of the change to make it easier to debug, review, and so that it's
smaller in case we have to roll it back. Unfortunately, that requires
duplicating some of the code in StandaloneTestStrategy until the next change
lands.

--
PiperOrigin-RevId: 145387109
MOS_MIGRATED_REVID=145387109