Adds Bazel support for Android NDK R13.
NDK13 changelog: https://github.com/android-ndk/ndk/wiki/Changelog-r13.

The most notable change is that the default compiler is now clang and GCC is not supported.

Also introduces an interface to encapsulate the creation and usage of ApiLevel and AndroidNdkCrosstoolsR* so that we can get rid of the repeated if-statements checking NDK major version.

RELNOTES: android_ndk_repository now supports Android NDK R13.

--
PiperOrigin-RevId: 145672022
MOS_MIGRATED_REVID=145672022