Bazel can now be built on Windows with MSVC
bazel build //src:bazel --cpu=x64_windows_msvc
now succeeds, hooray!

This change adds empty implementations for
build-runfiles and process-wrapper to make that
possible.

This means we can now build a bazel binary that
doesn't depend on MSYS.

The resulting binary is not yet functional because
many methods are still to be implemented, and they
just write "TODO: implement" or something similar.
But still this is great news, because now we can
add compile tests to the CI for MSVC!

See https://github.com/bazelbuild/bazel/issues/2107

--
PiperOrigin-RevId: 146106178
MOS_MIGRATED_REVID=146106178