allow skylark implicit output callbacks to use the rule name
Skylark implicit output callbacks could use any nonconfigurable
attribute of the rule except for "name". (As an implementation detail,
"name" is easy to overlook because it's a special case in rule
attribute map.) In practice, this isn't too much of a problem because
the return value of a skylark implicit output callback has %{}
substitutions applied to it, which do allow the rule name. However, it
is weird and inconsistent to prevent implicit output callbacks from
using the name.

--
Change-Id: I13149b2e9689ef8b8056c612a29df9da32e39bf3
Reviewed-on: https://cr.bazel.build/8251
PiperOrigin-RevId: 146178693
MOS_MIGRATED_REVID=146178693