ForgeClient accepts Spawn, update some callers
This is one step towards changing ForgeClient to require Spawn in all
cases. This is changing some of the easy ones. Also add a new Spawn
implementation class - for some reason BaseSpawn is trying to be helpful
and changing the environment around, as well as requiring that inputs
and outputs are inhertied from the owning action, which isn't always the
case. Ideally, BaseSpawn will go away in favor of the new
implementation, but that requires updating the callers that require the
special BaseSpawn functionality..

--
PiperOrigin-RevId: 146370998
MOS_MIGRATED_REVID=146370998